+++ collectors/plugins.d/pluginsd_functions.c
@@ -21,22 +21,23 @@ static void inflight_functions_insert_callback(const DICTIONARY_ITEM *item, void
     if(rc != 0)
         netdata_log_error("FUNCTION: '%s': cannot parse transaction UUID", string2str(pf->function));
 
-    bool has_payload = pf->payload && buffer_strlen(pf->payload) ? true : false;
-    char buffer[2048];
-
-    if(has_payload) {
-        snprintfz(
-            buffer, sizeof(buffer),
+    CLEAN_BUFFER *buffer = buffer_create(1024, NULL);
+    if(pf->payload && buffer_strlen(pf->payload)) {
+        buffer_sprintf(
+            buffer,
             PLUGINSD_KEYWORD_FUNCTION_PAYLOAD " %s %d \"%s\" \"%s\"\n",
             transaction,
             pf->timeout_s,
             string2str(pf->function),
             content_type_id2string(pf->payload->content_type)
             );
+
+        buffer_fast_strcat(buffer, buffer_tostring(pf->payload), buffer_strlen(pf->payload));
+        buffer_strcat(buffer, "\nFUNCTION_PAYLOAD_END\n");
     }
     else {
-        snprintfz(
-            buffer, sizeof(buffer),
+        buffer_sprintf(
+            buffer,
             PLUGINSD_KEYWORD_FUNCTION " %s %d \"%s\"\n",
             transaction,
             pf->timeout_s,
@@ -45,39 +46,25 @@ static void inflight_functions_insert_callback(const DICTIONARY_ITEM *item, void
     }
 
     // send the command to the plugin
-    ssize_t ret = send_to_plugin(buffer, parser);
-
+    // IMPORTANT: make sure all commands are sent in 1 call, because in streaming they may interfere with others
+    ssize_t ret = send_to_plugin(buffer_tostring(buffer), parser);
     pf->sent_monotonic_ut = now_monotonic_usec();
 
     if(ret < 0) {
+        pf->sent_successfully = false;
+
+        pf->code = HTTP_RESP_SERVICE_UNAVAILABLE;
         netdata_log_error("FUNCTION '%s': failed to send it to the plugin, error %zd", string2str(pf->function), ret);
-        rrd_call_function_error(pf->result_body_wb, "Failed to communicate with collector", HTTP_RESP_SERVICE_UNAVAILABLE);
+        rrd_call_function_error(pf->result_body_wb, "Failed to communicate with collector", pf->code);
     }
     else {
+        pf->sent_successfully = true;
+
         internal_error(LOG_FUNCTIONS,
                        "FUNCTION '%s' with transaction '%s' sent to collector (%zd bytes, in %"PRIu64" usec)",
                        string2str(pf->function), dictionary_acquired_item_name(item), ret,
                        pf->sent_monotonic_ut - pf->started_monotonic_ut);
     }
-
-    if (!has_payload)
-        return;
-
-    // send the payload to the plugin
-    ret = send_to_plugin(buffer_tostring(pf->payload), parser);
-
-    if(ret < 0) {
-        netdata_log_error("FUNCTION_PAYLOAD '%s': failed to send function to plugin, error %zd", string2str(pf->function), ret);
-        rrd_call_function_error(pf->result_body_wb, "Failed to communicate with collector", HTTP_RESP_SERVICE_UNAVAILABLE);
-    }
-    else {
-        internal_error(LOG_FUNCTIONS,
-                       "FUNCTION_PAYLOAD '%s' with transaction '%s' sent to collector (%zd bytes, in %"PRIu64" usec)",
-                       string2str(pf->function), dictionary_acquired_item_name(item), ret,
-                       pf->sent_monotonic_ut - pf->started_monotonic_ut);
-    }
-
-    send_to_plugin("\nFUNCTION_PAYLOAD_END\n", parser);
 }
 
 static bool inflight_functions_conflict_callback(const DICTIONARY_ITEM *item __maybe_unused, void *func __maybe_unused, void *new_func, void *parser_ptr __maybe_unused) {
@@ -222,7 +209,7 @@ int pluginsd_function_execute_cb(uuid_t *transaction, BUFFER *result_body_wb, BU
 
     usec_t now_ut = now_monotonic_usec();
 
-    int timeout_s = (*stop_monotonic_ut - now_ut + USEC_PER_SEC / 2) / USEC_PER_SEC;
+    int timeout_s = (int)((*stop_monotonic_ut - now_ut + USEC_PER_SEC / 2) / USEC_PER_SEC);
 
     struct inflight_function tmp = {
             .started_monotonic_ut = now_ut,
@@ -251,24 +238,34 @@ int pluginsd_function_execute_cb(uuid_t *transaction, BUFFER *result_body_wb, BU
 
     // if there is any error, our dictionary callbacks will call the caller callback to notify
     // the caller about the error - no need for error handling here.
-    void *t = dictionary_set(parser->inflight.functions, transaction_str, &tmp, sizeof(struct inflight_function));
-    if(register_canceller_cb)
-        register_canceller_cb(register_canceller_cb_data, pluginsd_function_cancel, t);
-
-    if(register_progresser_cb && (parser->repertoire == PARSER_INIT_PLUGINSD ||
-        (parser->repertoire == PARSER_INIT_STREAMING && stream_has_capability(&parser->user, STREAM_CAP_PROGRESS))))
-        register_progresser_cb(register_progresser_cb_data, pluginsd_function_progress_to_plugin, t);
-
-    if(!parser->inflight.smaller_monotonic_timeout_ut || *tmp.stop_monotonic_ut + RRDFUNCTIONS_TIMEOUT_EXTENSION_UT < parser->inflight.smaller_monotonic_timeout_ut)
-        parser->inflight.smaller_monotonic_timeout_ut = *tmp.stop_monotonic_ut + RRDFUNCTIONS_TIMEOUT_EXTENSION_UT;
-
-    // garbage collect stale inflight functions
-    if(parser->inflight.smaller_monotonic_timeout_ut < now_ut)
+    struct inflight_function *t = dictionary_set(parser->inflight.functions, transaction_str, &tmp, sizeof(struct inflight_function));
+    if(!t->sent_successfully) {
+        int code = t->code;
+        dictionary_del(parser->inflight.functions, transaction_str);
         pluginsd_inflight_functions_garbage_collect(parser, now_ut);
+        return code;
+    }
+    else {
+        if (register_canceller_cb)
+            register_canceller_cb(register_canceller_cb_data, pluginsd_function_cancel, t);
 
-    dictionary_write_unlock(parser->inflight.functions);
+        if (register_progresser_cb &&
+            (parser->repertoire == PARSER_INIT_PLUGINSD || (parser->repertoire == PARSER_INIT_STREAMING &&
+                                                            stream_has_capability(&parser->user, STREAM_CAP_PROGRESS))))
+            register_progresser_cb(register_progresser_cb_data, pluginsd_function_progress_to_plugin, t);
 
-    return HTTP_RESP_OK;
+        if (!parser->inflight.smaller_monotonic_timeout_ut ||
+            *tmp.stop_monotonic_ut + RRDFUNCTIONS_TIMEOUT_EXTENSION_UT < parser->inflight.smaller_monotonic_timeout_ut)
+            parser->inflight.smaller_monotonic_timeout_ut = *tmp.stop_monotonic_ut + RRDFUNCTIONS_TIMEOUT_EXTENSION_UT;
+
+        // garbage collect stale inflight functions
+        if (parser->inflight.smaller_monotonic_timeout_ut < now_ut)
+            pluginsd_inflight_functions_garbage_collect(parser, now_ut);
+
+        dictionary_write_unlock(parser->inflight.functions);
+
+        return HTTP_RESP_OK;
+    }
 }
 
 PARSER_RC pluginsd_function(char **words, size_t num_words, PARSER *parser) {
