+++ daemon/config/dyncfg-inline.c
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#define DYNCFG_INTERNALS
 #include "dyncfg.h"
 
 static DICTIONARY *dyncfg_nodes = NULL;
