+++ daemon/config/dyncfg.h
@@ -4,48 +4,8 @@
 #define NETDATA_DYNCFG_H
 
 #include "../common.h"
-
 #include "../../database/rrd.h"
 #include "../../database/rrdfunctions.h"
-
-#ifdef DYNCFG_INTERNALS
-typedef struct dyncfg {
-    RRDHOST *host;
-    uuid_t host_uuid;
-    STRING *path;
-    DYNCFG_STATUS status;
-    DYNCFG_TYPE type;
-    DYNCFG_CMDS cmds;
-    DYNCFG_SOURCE_TYPE source_type;
-    STRING *source;
-    usec_t created_ut;
-    usec_t modified_ut;
-    uint32_t saves;
-    bool sync;
-    bool user_disabled;
-    bool restart_required;
-
-    BUFFER *payload;
-
-    rrd_function_execute_cb_t execute_cb;
-    void *execute_cb_data;
-} DYNCFG;
-
-struct dyncfg_globals {
-    const char *dir;
-    DICTIONARY *nodes;
-};
-
-extern struct dyncfg_globals dyncfg_globals;
-
-void dyncfg_load_all(void);
-void dyncfg_load(const char *filename);
-void dyncfg_save(const char *id, DYNCFG *df);
-
-bool dyncfg_get_schema(const char *id, BUFFER *dst);
-
-void dyncfg_cleanup(DYNCFG *v);
-#endif
 
 void dyncfg_add_streaming(BUFFER *wb);
 bool dyncfg_available_for_rrdhost(RRDHOST *host);
